ai-dev-system/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
redis==6.0.0
pyyaml==6.0.2
requests==2.32.3
faiss-cpu==1.12.0
pytest==8.3.5
httpx==0.28.1
fakeredis==2.31.1
